:recycle: refactoring factory into multiple files
