+++ binding.gyp
@@ -7,7 +7,9 @@
         "src/bootstrap.c",
         "src/binding_util.c",
         "src/bridge.c",
-        "src/factory.c",
+        "src/factory/factory_common.c",
+        "src/factory/factory_regular.c",
+        "src/factory/factory_space.c",
         "src/convert_xyz.c",
         "src/convert_rgb.c",
         "src/convert_regular.c",
