+++ src/factory/factory_common.c
@@ -1,0 +1,18 @@
+#include "factory_common.h"
+#include "binding_error.h"
+#include "binding_util.h"
+
+napi_value BuildPromiseError(napi_env env, char *error) {
+    napi_status status;
+    napi_value object;
+    
+    status = napi_create_object(env, &object);
+    if (status != napi_ok) {
+        napi_throw_error(env, NULL, OBJ_MAKE_ERR);
+        return NULL;
+    }
+    
+    assignPropToJSObj(&object, env, string, "err", error);
+    
+    return object;
+}
